+++ src/components/calcite-date-month/calcite-date-month.tsx
@@ -42,6 +42,8 @@ export class CalciteDateMonth {
   @Prop() max: Date;
   /** User's language and region as BCP 47 formatted string. */
   @Prop() locale: string = "en-US";
+  /** specify the scale of the date picker */
+  @Prop({ reflect: true }) scale: "s" | "m" | "l";
 
   //--------------------------------------------------------------------------
   //
@@ -134,7 +136,10 @@ export class CalciteDateMonth {
     const month = this.activeDate.getMonth();
     const year = this.activeDate.getFullYear();
     const startOfWeek = getFirstDayOfWeek(this.locale);
-    const weekDays = getLocalizedWeekdays(this.locale);
+    const weekDays = getLocalizedWeekdays(
+      this.locale,
+      this.scale === "s" ? "narrow" : "short"
+    );
     const curMonDays = this.getCurrentMonthDays(month, year);
     const prevMonDays = this.getPrevMonthdays(month, year, startOfWeek);
     const nextMonDays = this.getNextMonthDays(month, year, startOfWeek);
@@ -143,6 +148,7 @@ export class CalciteDateMonth {
         const date = new Date(year, month - 1, day);
         return (
           <calcite-date-day
+            scale={this.scale}
             day={day}
             disabled={!inRange(date, this.min, this.max)}
             selected={sameDate(date, this.selectedDate)}
@@ -156,6 +162,7 @@ export class CalciteDateMonth {
         const active = sameDate(date, this.activeDate);
         return (
           <calcite-date-day
+            scale={this.scale}
             day={day}
             disabled={!inRange(date, this.min, this.max)}
             selected={sameDate(date, this.selectedDate)}
@@ -176,6 +183,7 @@ export class CalciteDateMonth {
         const date = new Date(year, month + 1, day);
         return (
           <calcite-date-day
+            scale={this.scale}
             day={day}
             disabled={!inRange(date, this.min, this.max)}
             selected={sameDate(date, this.selectedDate)}